LayoutTests/media/modern-media-controls/media-controller/media-controller-in-xhtml-expected.txt

Testing that media controls are correctly created in an XHTML document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.internals.shadowRoot(document.querySelector('video')).querySelector('.media-controls') instanceof HTMLDivElement is true
PASS successfullyParsed is true

TEST COMPLETE
